requirement.txt
torch
torchvision
transformers
datasets
streamlit
Pillow
scikit-learn
tqdm
